+++ src/test/java/org/folio/dew/EHoldingsTest.java
@@ -79,6 +79,7 @@ class EHoldingsTest extends BaseBatchTest {
   private final static String SINGLE_PACKAGE_ID = "1-23";
   private final static String PACKAGE_WITH_3_TITLES_ID = "1-21";
   private final static String PACKAGE_WITH_SAME_TITLE_NAMES_ID = "1-24";
+  private final static String PACKAGE_WITH_TITLE_AGREEMENTS_ID = "1-255";
   private final static String EXPECTED_RESOURCE_OUTPUT = "src/test/resources/output/eholdings_resource_export.csv";
   private final static String EXPECTED_PACKAGE_OUTPUT = "src/test/resources/output/eholdings_package_export.csv";
   private final static String EXPECTED_SINGLE_PACKAGE_OUTPUT =
@@ -87,6 +88,8 @@ class EHoldingsTest extends BaseBatchTest {
     "src/test/resources/output/eholdings_package_export_with_3_titles.csv";
   private final static String EXPECTED_PACKAGE_WITH_SAME_TITLE_NAMES_OUTPUT =
     "src/test/resources/output/eholdings_package_export_with_same_title_names.csv";
+  private final static String EXPECTED_PACKAGE_WITH_TITLE_AGREEMENTS_OUTPUT =
+    "src/test/resources/output/eholdings_package_export_with_title_agreements.csv";
   private static final String FILE_PATH = "mod-data-export-worker/e_holdings_export/diku/";
 
   @Test
@@ -158,6 +161,31 @@ class EHoldingsTest extends BaseBatchTest {
       getRequestedFor(
         urlEqualTo(
           "/eholdings/packages/1-21/resources?filter%5Bname%5D=%2A&sort=name&page=1&count=1")));
+
+    var packages = packageRepository.findAll();
+    var resources = resourceRepository.findAll();
+    assertEquals(0, ((Collection<?>) packages).size());
+    assertEquals(0, ((Collection<?>) resources).size());
+    verifyJobEvent();
+  }
+
+  @Test
+  @DisplayName("Run EHoldingsJob export package with title agreements successfully")
+  void eHoldingsJobPackageWithTitleAgreementsTest() throws Exception {
+    JobOperatorTestUtils testLauncher = createTestLauncher(getEHoldingsJob);
+    var exportConfig = buildExportConfig(PACKAGE_WITH_TITLE_AGREEMENTS_ID, PACKAGE);
+
+    final JobParameters jobParameters = prepareJobParameters(exportConfig);
+    JobExecution jobExecution = testLauncher.startJob(jobParameters);
+
+    verifyFile(jobExecution, EXPECTED_PACKAGE_WITH_TITLE_AGREEMENTS_OUTPUT);
+
+    assertThat(jobExecution.getExitStatus()).isEqualTo(ExitStatus.COMPLETED);
+
+    wireMockServer.verify(
+      getRequestedFor(
+        urlEqualTo(
+          "/eholdings/packages/1-255/resources?filter%5Bname%5D=%2A&sort=name&page=1&count=1")));
 
     var packages = packageRepository.findAll();
     var resources = resourceRepository.findAll();
@@ -277,6 +305,10 @@ class EHoldingsTest extends BaseBatchTest {
     final String presignedUrl = remoteFilesStorage.objectToPresignedObjectUrl(fileInStorage);
     final FileSystemResource actualOutput = actualFileOutput(presignedUrl);
     FileSystemResource expectedOutput = new FileSystemResource(expectedFile);
+    if (!FileUtils.contentEqualsIgnoreEOL(expectedOutput.getFile(), actualOutput.getFile(), "UTF-8")) {
+      log.error("ACTUAL OUTPUT:\n{}", FileUtils.readFileToString(actualOutput.getFile(), "UTF-8"));
+      log.error("EXPECTED OUTPUT:\n{}", FileUtils.readFileToString(expectedOutput.getFile(), "UTF-8"));
+    }
     assertTrue(FileUtils.contentEqualsIgnoreEOL(expectedOutput.getFile(), actualOutput.getFile(), "UTF-8"), "Files are not identical!");
   }